Refactor SnapshotVersions into its own file.
