+++ src/screens/internal/CompareSnapshots/CompareSnapshots.tsx
@@ -1,13 +1,10 @@
 import React, { Fragment, useEffect, useState } from 'react';
-import Grid from '@material-ui/core/Grid';
 import { Wrapper } from './CompareSnapshots.style';
-import { SnapshotVersion, Api } from 'api';
-import moment from 'moment';
-import { snapshotUrl } from 'common/utils/snapshots';
 import { reenableDisabledMetrics } from 'common/models/Projection';
 import { ComparisonList } from './ComparisonList';
 import { CompareOptions, useProjectionsSet } from './utils';
 import { OptionsSelector } from './OptionsSelector';
+import { SnapshotVersions } from './SnapshotVersions';
 
 export function CompareSnapshots() {
   // We want to force all metrics to be reenabled so we can evaluate whether they're fixed.
@@ -36,22 +33,12 @@ function CompareSnapshotsBody({ options }: { options: CompareOptions }) {
   );
   projectionsSet = projectionsSet.sortBy(sortType, metric);
 
-  const leftVersion = useSnapshotVersion(leftSnapshot);
-  const rightVersion = useSnapshotVersion(rightSnapshot);
-
   return (
     <Fragment>
-      <Grid container spacing={8} style={{ margin: '1px' }}>
-        <Grid item xs={6}>
-          Left Snapshot: <b>{leftSnapshot}</b>
-          <VersionInfo version={leftVersion} />
-        </Grid>
-        <Grid item xs={6}>
-          Right Snapshot: <b>{rightSnapshot}</b>
-          <VersionInfo version={rightVersion} />
-        </Grid>
-      </Grid>
-
+      <SnapshotVersions
+        leftSnapshot={leftSnapshot}
+        rightSnapshot={rightSnapshot}
+      />
       <ComparisonList
         metric={metric}
         projectionsSet={projectionsSet}
@@ -59,44 +46,6 @@ function CompareSnapshotsBody({ options }: { options: CompareOptions }) {
       />
     </Fragment>
   );
-}
-
-const VersionInfo = function ({
-  version,
-}: {
-  version: SnapshotVersion | null;
-}) {
-  return (
-    version && (
-      <div style={{ fontSize: 'small' }}>
-        <b>Build finished:</b>{' '}
-        {moment.utc(version.timestamp).local().toDate().toString()}
-        <br />
-        <b>covid-data-model:</b>{' '}
-        {JSON.stringify(version['covid-data-model']).replace(',', ', ')}
-        <br />
-        <b>covid-data-public:</b>{' '}
-        {JSON.stringify(version['covid-data-public']).replace(',', ', ')}
-        <br />
-      </div>
-    )
-  );
-};
-
-export function useSnapshotVersion(
-  snapshot: number | null,
-): SnapshotVersion | null {
-  const [version, setVersion] = useState<SnapshotVersion | null>(null);
-  useEffect(() => {
-    setVersion(null);
-    if (snapshot !== null) {
-      new Api(snapshotUrl(snapshot)).fetchVersionInfo().then(version => {
-        setVersion(version);
-      });
-    }
-  }, [snapshot]);
-
-  return version;
 }
 
 export default CompareSnapshots;
